feat(205-go-app): use env variables to improve GC performance (#265)
Go has 2 variables to control it's garbage collection behaviour, and
ultimately trade off memory to spend less time in the GC.

The GC settings are exposed as both environment variables but can also be
set in the code, although the former makes more sense for values that
could change based on the environment (e.g. set GOMEMLIMIT based on
container size).

GOGC: controls the GC threshold by serving as a multiplier of the active
memory after the previous collection. 300 (up from 100) means that the next
threshold will be set to (100+GOGC=400%) of the active memory found during
the last collection. The value of 300 means 3x as much memory will be
allocated, on average, before the next GC, compared to the default value.

GOMEMLIMIT: controls the GC threshold via a static (soft) limit. The final
limit is essentially the minimum of the two calculations, and this limit
allows one to set GOGC quite high (or off completly) while still keeping
memory use constrained. The limit is soft in that "Garbage" will be
collected to stay under the limit, but if actual in-use memory exceeds the
limit, the program will GC thrash since it can't get below it.

However provided the limit is appropriate, it's more likely your
program is mismanaging it's memory, so not having the limit would have
resulted in swapping or OOM-ing. The default limit is infinity, as it is
the newer control, and this provides identical behaviour to programs
before this new limit type was added.

I set the value arbitrarily to 200 MiB, as I am not sure what else is
running in the container with it so I left a bunch of room. It should
ensure that the Go program will never set it's GC limit above that,
although unless there is more than 50 MiB of active memory after a
collection the current value of GOGC will not produce a target value that
high anyway.

For the simple test, using wrk's default (-d10 -t2 -c10), and setting
GODEBUG=gctrace=1 to see GC stats:

wrk -d10 -t2 -c10:
	GOGC=100 - 1044 GC Cycles
	  857696 requests in 10.01s, 470.33MB read
	GOGC=300 - 334 GC Cycles
	  1150434 requests in 10.00s, 630.86MB read
34% more requests, 68% less GC cycles

wrk -d30 -t12 -c400:
	GOGC=100 - 1711 GC Cycles
	  2766113 requests in 30.10s, 1.48GB read
	GOGC=300 - 665 GC Cycles
	  3598185 requests in 30.09s, 1.93GB read
30% more requests, 61% less GC cycles

The active memory never exceeded 4-5 MiB, so the computed limits never got
higher than 22 MiB, meaning GOMEMLIMIT would not come into effect.