refactor(205/go-app): use io.WriteString
When writing strings directly to an io.Writer, the preferred method (if
not using fmt.Fprint*) is to use io.WriteString, a helper that accepts
the writer and the string.

On the surface this basically saves you from making the conversion to a
byte slice yourself resulting in cleaner code, but it also has a
performance benefit, since it prevents a bunch of extra string->[]byte
conversions, since it checks if the underlying writer has a WriteString
method to call instead.

The main goal isn't the performance optimization though, as none of the
strings involved are that large to really make a difference. This is
about making everything cleaner and simpler to use.