fix(205/go-app): pipe contexts through code
Update functions and methods that eventually call context requiring code
to themselves accept a context. This will let us get rid of the calls to
create new contexts and the TODO placeholders as we can pass a single
context through to each.

Also "create" appropriate contexts for the two general cases they are
used:
 - The startup code generates a single context to do the initial setup
   of the DB and S3 services.
 - The http handlers grab the context from their request and pass it to
   the helper code to handle the S3 uploads and DB updates.

This will make it easier to set up timeouts, as well as parallelize
sub-tasks to ensure cancellation / error signals propgate properly. In
the case of the HTTP handlers it already allows the S3 uploads / DB
updates to be cancelled if the http request is cancelled by the client.