refactor(205/go-app): split up server setup
Move the config loading and prometheus setup outside the setup
"constructor" for the main server. All that's left in the "constructor"
is setting up the prometheus metrics (which don't need a server to be
running, and setting up the S3/DB connections. Those latter options are
lazily performed, so if we don't actually hit the respective endpoints
no attempt to connect to Postgres/S3 are made. I'm not going to the
effort to create fakes for those (yet...), but they're safe to leave
as-is.

These changes allow the base (non-external) functionality of the server
to be tested and benchmarked. The additional features the server
normally has are now the responsibility of main and helper functions
with the common elements (e.g. the config and prometheus registry)
factored out as parameters.