feat(205/go-app): add tests of base server + PGO
Now that it's possible to run the server code safely for some endpoints
in tests, create tests and benchmarks for the two currently "safe"
endpoints: /healthz and /api/devices

The tests are fairly simple, as both endpoints are expected to return
known values every time, but will allow for refactoring/optimization
with confidence.

The benchmarks perform a full HTTP round-trip stressing both client and
server. These are not meant to be representative of real-world traffic
and memory use, but can still be useeful to compare code.

They also allow for the easy creation of CPU (and Memory) profiles for
these cases to help find places to optimize manually. In addition the
CPU profile can be fed to the Go compiler to enable Profile Guided
Optimizations to be performed. PGO allows for deeper inlining, as well
as interface devirtualization, and a growing list of other
optimizations.

Note: PGO is not automatically enabled, nor is a "default" profile being
included in this change to do so, as this change like its predecessors
is not focused on modifying behaviour or making optimizations.

However if you want to try, simply run `go generate` to create a sample
"default.pgo" file in the currently directory to enable PGO in local and
container builds.