feat(205/go-app): add clean benchmark
The original benchmark for the endpoints was fine for PGO purposes as it
captures the full process of decoding HTTP requests as performed by the
server. The client side of the benchmark although not useful for
building the final server, is simply ignored by the compiler.

However, when looking at the benchmark results and analyzing the
profiles manually, this adds a lot of unneded noise that makes it harder
to see due to much of the memory allocated during the benchmark is
actually for client-side objects that are thrown away on every
iteration.

Create another benchmark that targets more specifically the endpoint
code itself by simply using the Mux/router to find and call the endpoint
using httptest Mock objects. This benchmark thus represents the part of
server the user is more directly responsible for, but the limited scope
and use of mock objects makes these benchmarks less useful for PGO on
their own. The stats of the benchmark (cpu and memory use) although not
representative of real world scenarios, should be even more useful to
analyze the results of code changes.

The go generate command will incorporate both these new benchmarks and
the older "EndToEnd" ones, as the new ones still let us tell the
compiler that we are very interested in optimizing the endpoint handlers
themselves.

Also, while we're here, extract the leading "/" from the endpoint name
when using it to create the sub-benchmark names, as before it resulted
in lines like "BenchmarkEndpoints//api/devices..."