feat(205/go-app): parallelize getImage parts
Like with the change to the Rust code in #258, the two actions of the getImages handler can be done in parallel.

In this case we can use the golang.org/x/sync/errorgroup package to run both tasks in parallel catching any errors. By propogating the same (errgrp generated) context to the tasks, both tasks will be sent a cancel signal if either of them fail.

Technically the behaviour of the code has changed, since now we could fail to upload to S3 but still write an entry to the database (if that happens first), but that is the same behaviour that the Rust change introduced.

I believe the Rust code is guaranteed to write to the DB regardless of the result of the S3 upload, and need to wait for both to finish, wheras Go can fail faster and a quick S3 failure will prevent a database update.