Go cleanups for tutorial 205 (#266)
* refactor(205/go-app): use io.WriteString

When writing strings directly to an io.Writer, the preferred method (if
not using fmt.Fprint*) is to use io.WriteString, a helper that accepts
the writer and the string.

On the surface this basically saves you from making the conversion to a
byte slice yourself resulting in cleaner code, but it also has a
performance benefit, since it prevents a bunch of extra string->[]byte
conversions, since it checks if the underlying writer has a WriteString
method to call instead.

The main goal isn't the performance optimization though, as none of the
strings involved are that large to really make a difference. This is
about making everything cleaner and simpler to use.

* fix(205/go-app): pipe contexts through code

Update functions and methods that eventually call context requiring code
to themselves accept a context. This will let us get rid of the calls to
create new contexts and the TODO placeholders as we can pass a single
context through to each.

Also "create" appropriate contexts for the two general cases they are
used:
 - The startup code generates a single context to do the initial setup
   of the DB and S3 services.
 - The http handlers grab the context from their request and pass it to
   the helper code to handle the S3 uploads and DB updates.

This will make it easier to set up timeouts, as well as parallelize
sub-tasks to ensure cancellation / error signals propgate properly. In
the case of the HTTP handlers it already allows the S3 uploads / DB
updates to be cancelled if the http request is cancelled by the client.

* refactor(205/go-app): split up server setup

Move the config loading and prometheus setup outside the setup
"constructor" for the main server. All that's left in the "constructor"
is setting up the prometheus metrics (which don't need a server to be
running, and setting up the S3/DB connections. Those latter options are
lazily performed, so if we don't actually hit the respective endpoints
no attempt to connect to Postgres/S3 are made. I'm not going to the
effort to create fakes for those (yet...), but they're safe to leave
as-is.

These changes allow the base (non-external) functionality of the server
to be tested and benchmarked. The additional features the server
normally has are now the responsibility of main and helper functions
with the common elements (e.g. the config and prometheus registry)
factored out as parameters.

* feat(205/go-app): add tests of base server + PGO

Now that it's possible to run the server code safely for some endpoints
in tests, create tests and benchmarks for the two currently "safe"
endpoints: /healthz and /api/devices

The tests are fairly simple, as both endpoints are expected to return
known values every time, but will allow for refactoring/optimization
with confidence.

The benchmarks perform a full HTTP round-trip stressing both client and
server. These are not meant to be representative of real-world traffic
and memory use, but can still be useeful to compare code.

They also allow for the easy creation of CPU (and Memory) profiles for
these cases to help find places to optimize manually. In addition the
CPU profile can be fed to the Go compiler to enable Profile Guided
Optimizations to be performed. PGO allows for deeper inlining, as well
as interface devirtualization, and a growing list of other
optimizations.

Note: PGO is not automatically enabled, nor is a "default" profile being
included in this change to do so, as this change like its predecessors
is not focused on modifying behaviour or making optimizations.

However if you want to try, simply run `go generate` to create a sample
"default.pgo" file in the currently directory to enable PGO in local and
container builds.

* feat(205/go-app): add clean benchmark

The original benchmark for the endpoints was fine for PGO purposes as it
captures the full process of decoding HTTP requests as performed by the
server. The client side of the benchmark although not useful for
building the final server, is simply ignored by the compiler.

However, when looking at the benchmark results and analyzing the
profiles manually, this adds a lot of unneded noise that makes it harder
to see due to much of the memory allocated during the benchmark is
actually for client-side objects that are thrown away on every
iteration.

Create another benchmark that targets more specifically the endpoint
code itself by simply using the Mux/router to find and call the endpoint
using httptest Mock objects. This benchmark thus represents the part of
server the user is more directly responsible for, but the limited scope
and use of mock objects makes these benchmarks less useful for PGO on
their own. The stats of the benchmark (cpu and memory use) although not
representative of real world scenarios, should be even more useful to
analyze the results of code changes.

The go generate command will incorporate both these new benchmarks and
the older "EndToEnd" ones, as the new ones still let us tell the
compiler that we are very interested in optimizing the endpoint handlers
themselves.

Also, while we're here, extract the leading "/" from the endpoint name
when using it to create the sub-benchmark names, as before it resulted
in lines like "BenchmarkEndpoints//api/devices..."