feat(205/go-app): write JSON directly to response
Update the renderJSON function to write the encoded JSON data directly to
the writer instead of getting the bytes manually. This cuts down on both
an allocation and a copy, which is both faster and allocates less. The
bigger benefit is the latter, as aside from the cost of allocation itself,
allocating less means more time before the next GC, and likely less
"live" memory, helping out when the server is under load to the point that
GC could delay responses.

Putting the header assignment before everything is safe, as the header
is only written once bytes are written to the output, and the error case
actually re-sets the content-type when writing out the error.

The json.Encoder writing data to the response directly actually has a
buffer in between, so it doesn't write anything should the JSON encoding
fail. However, unlike the original code, the buffer it uses is pooled
and thus the memory is re-used across requests instead of becoming
garbage.

The only "noticeable" difference is that the resulting JSON when using
the encoder to write to the result adds a newline at the end of the
output. This is because the encoder also expects to be used to stream
multiple JSON values in succession, so the newline is there to ensure no
issues when a value to be encoded is just a number.

Benchmark results:
```
                         │ bench-base.txt │           bench-json.txt             │
                         │     sec/op     │   sec/op     vs base                 │
EndToEnd/api/devices-32       271.2m ± 3%   278.6m ± 2%  +2.74% (p=0.004 n=10)
Endpoints/api/devices-32      3.379m ± 2%   3.091m ± 2%  -8.53% (p=0.000 n=10)

                         │ bench-base.txt │            bench-json.txt            │
                         │      B/op      │     B/op      vs base                │
EndToEnd/api/devices-32      18.37Mi ± 0%   17.89Mi ± 0%   -2.61% (p=0.000 n=10)
Endpoints/api/devices-32     2.072Mi ± 0%   1.613Mi ± 0%  -22.15% (p=0.000 n=10)

                         │ bench-base.txt │            bench-json.txt            │
                         │   allocs/op    │  allocs/op   vs base                 │
EndToEnd/api/devices-32       131.1k ± 0%   130.1k ± 0%  -0.78% (p=0.000 n=10)
Endpoints/api/devices-32      12.01k ± 0%   11.01k ± 0%  -8.35% (p=0.000 n=10)
```

That one allocation was nearly a quarter of the total memory allocated
by the handler. This results in an 8-9% performance improvement in both
latency and throughput on wrk based tests.

wrk -t2 -c10 -d30s: (Light Load)
  Thread Stats   Avg      Stdev     Max   +/- Stdev
Before:
    Latency   163.41us  129.71us   1.71ms   82.72%
    Req/Sec    33.76k     1.76k   40.56k    68.44%
After:
    Latency   149.51us  121.89us   3.86ms   83.67%
    Req/Sec    37.13k     1.79k   46.74k    75.75%
wrk -t12 -c200 -d30s: (Heavy Load)
Before:
    Latency     2.07ms    2.80ms  47.44ms   89.16%
    Req/Sec    23.79k     1.92k   41.22k    69.83%
After:
    Latency     1.90ms    2.54ms  44.58ms   88.70%
    Req/Sec    25.65k     2.11k   43.12k    70.36%