fix(205/go-app): make benchmarks better
There were some notable issues with the benchmarks:
 - turning on profiling sped them up due to the interaction of parallel
   benchmarking and the profiling causing the GC to run less frequently
 - the allocations for a single iteration don't accurately represent the
   allocations when memory-pooling is in place, like in the HTTP server
 - I generally don't understand the parallel benchmark feature, and
   should have probably avoided it in the first place :)
 - trying to use the benchmark to compare before/after profiles didn't
   make sense since faster code would be run more times.

I've updated the benchmarks instead to be single-threaded, and have a
fixed size "work-unit" of 1000 requests. This should help the benchmark
work better, represent the effect of memory pooling better, and allow
for a fixed amount of work to be done when looking at profiles manually.