fix(205/go-app): fail getImage better
Introduce a db transaction to wrap most of getImage so that if anything goes wrong, the database transaction will be rolled-back.

This means that any failure to upload to S3 will again prevent the DB from being updated.

This doesn't guarantee that the S3 upload will be rolled-back should it complete first. But if the DB should fail before it the s3 upload will have it's context cancelled which may cancel the upload.

This is still better than both the initial parallel implementation (including the Rust counterpart) and the non-parallel version because there it could easily be the case that S3 completes successfully and then the DB fails, but it never looped back to "undo" the upload. This way, the s3 request might get cancelled if the DB fails faster.