refactor(205/go-app): cleanup errors and metrics
Add a helper "annotate" which does 2 things:
 - it does error wrapping to add context in a more consistent way.
 - it does nothing if there is no error, simplifying some cases for callers.

This cleans up a few cases including where the final error return is inside an annoying "if err != nil" block, instead allowing us to return the result of annotate since it will pass through a nil error.

Also, in the two helpers that assign prometheus metrics, do the assignment via a deferred function to make it not interfere with exiting the helper. To maintain similarity with the original code, the final error for the helper is checked before submitting the prometheus metrics, although even in the original non-parallel code a metric may or may not have been updated due to how the getImage request failed.