fix(205/go-app): fix broken db insert
The improvements in PR #256 actually broke the DB insert. It was trying to hard-code the table into the query (previously the table was sptrintf'd in) and escape it to prevent any shenanigans.

Unfortunately backticks "`" are not valid to surround a table name in the query string. Single quotes denote text in a SQL string, and double quotes can mean to escape a name such that a table or column name can contain special characters.

Switch it up by using backticks in the Go source to turn the query string into a raw string, and then use double-quotes inside the string to quote the table without needed to escape them.