Go parallel getImage handler, better errors, and fix DB (#279)
* feat(205/go-app): parallelize getImage parts

Like with the change to the Rust code in #258, the two actions of the getImages handler can be done in parallel.

In this case we can use the golang.org/x/sync/errorgroup package to run both tasks in parallel catching any errors. By propogating the same (errgrp generated) context to the tasks, both tasks will be sent a cancel signal if either of them fail.

Technically the behaviour of the code has changed, since now we could fail to upload to S3 but still write an entry to the database (if that happens first), but that is the same behaviour that the Rust change introduced.

I believe the Rust code is guaranteed to write to the DB regardless of the result of the S3 upload, and need to wait for both to finish, wheras Go can fail faster and a quick S3 failure will prevent a database update.

* fix(205/go-app): fail getImage better

Introduce a db transaction to wrap most of getImage so that if anything goes wrong, the database transaction will be rolled-back.

This means that any failure to upload to S3 will again prevent the DB from being updated.

This doesn't guarantee that the S3 upload will be rolled-back should it complete first. But if the DB should fail before it the s3 upload will have it's context cancelled which may cancel the upload.

This is still better than both the initial parallel implementation (including the Rust counterpart) and the non-parallel version because there it could easily be the case that S3 completes successfully and then the DB fails, but it never looped back to "undo" the upload. This way, the s3 request might get cancelled if the DB fails faster.

* refactor(205/go-app): cleanup errors and metrics

Add a helper "annotate" which does 2 things:
 - it does error wrapping to add context in a more consistent way.
 - it does nothing if there is no error, simplifying some cases for callers.

This cleans up a few cases including where the final error return is inside an annoying "if err != nil" block, instead allowing us to return the result of annotate since it will pass through a nil error.

Also, in the two helpers that assign prometheus metrics, do the assignment via a deferred function to make it not interfere with exiting the helper. To maintain similarity with the original code, the final error for the helper is checked before submitting the prometheus metrics, although even in the original non-parallel code a metric may or may not have been updated due to how the getImage request failed.

* fix(205/go-app): fix broken db insert

The improvements in PR #256 actually broke the DB insert. It was trying to hard-code the table into the query (previously the table was sptrintf'd in) and escape it to prevent any shenanigans.

Unfortunately backticks "`" are not valid to surround a table name in the query string. Single quotes denote text in a SQL string, and double quotes can mean to escape a name such that a table or column name can contain special characters.

Switch it up by using backticks in the Go source to turn the query string into a raw string, and then use double-quotes inside the string to quote the table without needed to escape them.